requirements.txt
asgiref==3.3.4
Django==3.2.4
django-crispy-forms==1.12.0
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
